requirements.txt
cbpro 
numpy 
pymongo 
PyYAML 
requests 
six 
sortedcontainers 
TA-Lib 
websocket-client 
